feat: HH : mm 시간 타입 처리 관련 extension 추가
